refactor(core): move exceedsMax out of scale.js into its own util
scale.js helpers all *produce* a max (western/myriad/indian/longScale/
bounded/UNBOUNDED); exceedsMax *consumes* one — a range check, not a scale
derivation. Move it to src/utils/exceeds-max.js (single-purpose, matching the
repo's util convention) and repoint the 18 migrated languages' imports.

Behaviour identical; scale.js now produces maxima only, exceeds-max.js checks
them, too-large-error.js reports the failure. npm test: 362 passing; lint clean.
